+++ api/src/__tests__/contract.test.js
@@ -25,8 +25,8 @@ jest.mock("node-fetch", () =>
   ),
 );
 
-jest.mock("../sendinblue", () => ({
-  ...jest.requireActual("../sendinblue"),
+jest.mock("../brevo", () => ({
+  ...jest.requireActual("../brevo"),
   sendEmail: () => Promise.resolve(),
 }));
