+++ api/src/__tests__/young.test.ts
@@ -44,8 +44,8 @@ jest.mock("../redis", () => {
   };
 });
 
-jest.mock("../sendinblue", () => ({
-  ...jest.requireActual("../sendinblue"),
+jest.mock("../brevo", () => ({
+  ...jest.requireActual("../brevo"),
   sendEmail: () => Promise.resolve(),
 }));
