+++ api/src/controllers/SNUpport.js
@@ -13,7 +13,7 @@ const { capture } = require("../sentry");
 const SNUpport = require("../SNUpport");
 const { ERRORS, isYoung, uploadFile, getFile, SUPPORT_BUCKET_CONFIG } = require("../utils");
 const config = require("config");
-const { sendTemplate } = require("../sendinblue");
+const { sendTemplate } = require("../brevo");
 const ReferentObject = require("../models/referent");
 const YoungObject = require("../models/young");
 const ClasseObject = require("../models/cle/classe");
