+++ api/src/controllers/cle/referent-signup.js
@@ -8,7 +8,7 @@ const emailsEmitter = require("../../emails");
 const config = require("config");
 const { capture } = require("../../sentry");
 const { ERRORS, validatePassword } = require("../../utils");
-const { sendTemplate } = require("../../sendinblue");
+const { sendTemplate } = require("../../brevo");
 const ReferentModel = require("../../models/referent");
 const EtablissementModel = require("../../models/cle/etablissement");
 const ClasseModel = require("../../models/cle/classe");
