+++ api/src/controllers/email.js
@@ -11,7 +11,7 @@ const { capture, captureMessage } = require("../sentry");
 const EmailObject = require("../models/email");
 const { canViewEmailHistory } = require("snu-lib");
 const { serializeEmail } = require("../utils/serializer");
-const { getEmailsList, getEmailContent } = require("../sendinblue");
+const { getEmailsList, getEmailContent } = require("../brevo");
 const { validateId } = require("../utils/validator");
 
 router.get("/", passport.authenticate(["referent"], { session: false, failWithError: true }), async (req, res) => {
