+++ api/src/controllers/waiting-list.js
@@ -5,7 +5,7 @@ const { capture } = require("../sentry");
 const WaitingListModel = require("../models/waitingList");
 const { ERRORS } = require("../utils");
 const { validateWaitingList } = require("../utils/validator");
-const { sendTemplate } = require("../sendinblue");
+const { sendTemplate } = require("../brevo");
 const { SENDINBLUE_TEMPLATES } = require("snu-lib");
 
 router.post("/", async (req, res) => {
