+++ api/src/crons/missionOutdated.js
@@ -4,7 +4,7 @@ const Mission = require("../models/mission");
 const Referent = require("../models/referent");
 const ApplicationObject = require("../models/application");
 const YoungObject = require("../models/young");
-const { sendTemplate } = require("../sendinblue");
+const { sendTemplate } = require("../brevo");
 const slack = require("../slack");
 const { SENDINBLUE_TEMPLATES, APPLICATION_STATUS } = require("snu-lib");
 const config = require("config");
