+++ api/src/emails/cle/referent.js
@@ -2,7 +2,7 @@ const { SENDINBLUE_TEMPLATES } = require("snu-lib");
 const EtablissementModel = require("../../models/cle/etablissement");
 const ClasseModel = require("../../models/cle/classe");
 const { capture } = require("../../sentry");
-const { sendTemplate } = require("../../sendinblue");
+const { sendTemplate } = require("../../brevo");
 
 module.exports = (emailsEmitter) => {
   // Coordinator
