+++ api/src/services/application.js
@@ -4,7 +4,7 @@ const ApplicationModel = require("../models/application");
 const CohortModel = require("../models/cohort");
 const YoungModel = require("../models/young");
 const ReferentModel = require("../models/referent");
-const { sendTemplate } = require("../sendinblue");
+const { sendTemplate } = require("../brevo");
 const config = require("config");
 const { getCcOfYoung } = require("../utils");
 const { getTutorName } = require("./mission");
