+++ api/src/utils/index.js
@@ -12,7 +12,7 @@ const ReferentModel = require("../models/referent");
 const ContractObject = require("../models/contract");
 const SessionPhase1 = require("../models/sessionPhase1");
 const CohortModel = require("../models/cohort");
-const { sendEmail, sendTemplate } = require("../sendinblue");
+const { sendEmail, sendTemplate } = require("../brevo");
 const path = require("path");
 const fs = require("fs");
 const { addDays } = require("date-fns");
